pyproject.toml
[tool.poetry]
name = "gen-ai-playground"
version = "0.1.0"
description = ""
authors = ["Abel"]
license = "MIT"
readme = "README.md"
packages = [{include = "gen_ai_playground"}]

[tool.poetry.dependencies]
python = "^3.10"
sentence-transformers = "^2.2.2"
langchain = "^0.0.187"
openai = "^0.27.7"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
